Fix migration

+++ migrations/sqls/20230709035653-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY NOT NULL,
     name VARCHAR(255) NOT NULL,
-    price NUMERIC NOT NULL,
+    price integer NOT NULL,
     category VARCHAR(50)
 );
